added updated classes, program.cs and readme

+++ Policies/IPenaltyCalculator.cs
@@ -2,6 +2,6 @@ namespace APDB_Cw1_s21417.Policies;
 
 public interface IPenaltyCalculator
 {
-    double Calculate(DateTime dueDate, DateTime returnDate);
+    decimal Calculate(DateTime dueDate, DateTime returnDate);
 }
